K-Nearest Neighbors - Skill Lab
After completing this lab you will be able to:

Use K Nearest neighbors to classify data

K-Nearest Neighbors is a supervised learning algorithm. Where the data is 'trained' with data points corresponding to their classification. To predict the class of a given data point, it takes into account the classes of the 'K' nearest data points and chooses the class in which the majority of the 'K' nearest data points belong to as the predicted class.